requirements.txt
acrome-smd==1.1.4
build
ipython
